Added a EraseAccount() and RunDebug() method , adjusted user class to include a way to display their stats.
EraseAccount() removes member,
RunDebug() is accessed by us to check info about all the users (for now), the password is "flyhigh" by the way.

Fixed the Id numbers given to the players so they get avaialble numbers if someone with an earlier number was deleted.

Added some extra layers of polish overall to the AccountManager, some colors to the menu too.